+++ src/Jobs/Reservers/DefaultReserver.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Qless\Jobs\Reservers;
+
+class DefaultReserver extends AbstractReserver implements ReserverInterface
+{
+    /**
+     * {@inheritdoc}
+     *
+     * @var string
+     */
+    const TYPE_DESCRIPTION = 'default';
+}