+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Added `DefaultReserver` which works just like `OrderedReserver` with one exception: it does not sort the queues
 
+### Fixed
+- Fixed  `OrderedReserver` to sort queues using natural ordering like [`natsort()`](http://php.net/manual/en/function.natsort.php)
+
 ## [3.0.0] - 2018-10-19
 ### Changed
 - Brand new Events Model: